+++ R/expr__datetime.R
@@ -35,7 +35,7 @@
 #' df
 ExprDT_truncate = function(every, offset = NULL) {
   .pr$Expr$dt_truncate(self, every, offset) |>
-    unwrap("in dt$truncate()")
+    unwrap("in $dt$truncate()")
 }
 
 #' Round datetime
@@ -83,7 +83,7 @@ ExprDT_truncate = function(every, offset = NULL) {
 #' df
 ExprDT_round = function(every, offset = NULL) {
   .pr$Expr$dt_round(self, every, offset) |>
-    unwrap("in dt$round()")
+    unwrap("in $dt$round()")
 }
 
 # ExprDT_combine = function(self, tm: time | pli.RPolarsExpr, tu: TimeUnit = "us") -> pli.RPolarsExpr:
@@ -547,7 +547,7 @@ ExprDT_epoch = function(tu = c("us", "ns", "ms", "s", "d")) {
     or_else = Err(
       paste("tu must be one of 'ns', 'us', 'ms', 's', 'd', got", str_string(tu))
     )
-  ) |> map_err(\(err) paste("in dt$epoch:", err))
+  ) |> map_err(\(err) paste("in $dt$epoch:", err))
 
   unwrap(expr_result)
 }
@@ -575,7 +575,7 @@ ExprDT_epoch = function(tu = c("us", "ns", "ms", "s", "d")) {
 #' )
 ExprDT_timestamp = function(tu = c("ns", "us", "ms")) {
   .pr$Expr$timestamp(self, tu[1]) |>
-    map_err(\(err) paste("in dt$timestamp:", err)) |>
+    map_err(\(err) paste("in $dt$timestamp:", err)) |>
     unwrap()
 }
 
@@ -604,7 +604,7 @@ ExprDT_timestamp = function(tu = c("ns", "us", "ms")) {
 #' )
 ExprDT_with_time_unit = function(tu = c("ns", "us", "ms")) {
   .pr$Expr$dt_with_time_unit(self, tu[1]) |>
-    map_err(\(err) paste("in dt$with_time_unit:", err)) |>
+    map_err(\(err) paste("in $dt$with_time_unit:", err)) |>
     unwrap()
 }
 
@@ -634,7 +634,7 @@ ExprDT_with_time_unit = function(tu = c("ns", "us", "ms")) {
 #' )
 ExprDT_cast_time_unit = function(tu = c("ns", "us", "ms")) {
   .pr$Expr$dt_cast_time_unit(self, tu[1]) |>
-    map_err(\(err) paste("in dt$cast_time_unit:", err)) |>
+    map_err(\(err) paste("in $dt$cast_time_unit:", err)) |>
     unwrap()
 }
 
@@ -661,7 +661,7 @@ ExprDT_cast_time_unit = function(tu = c("ns", "us", "ms")) {
 ExprDT_convert_time_zone = function(time_zone) {
   check_tz_to_result(time_zone) |>
     and_then(\(valid_tz) .pr$Expr$dt_convert_time_zone(self, valid_tz)) |>
-    map_err(\(err) paste("in dt$convert_time_zone:", err)) |>
+    map_err(\(err) paste("in $dt$convert_time_zone:", err)) |>
     unwrap("in $convert_time_zone():")
 }
 
@@ -947,5 +947,19 @@ ExprDT_offset_by = function(by) {
 #' df$with_columns(times = pl$col("dates")$dt$time())
 ExprDT_time = function() {
   .pr$Expr$dt_time(self) |>
-    unwrap("in dt$time()")
+    unwrap("in $dt$time()")
+}
+
+#' Determine whether the year of the underlying date is a leap year
+#'
+#' @return An Expr of datatype Bool
+#'
+#' @examples
+#' df = pl$DataFrame(date = as.Date(c("2000-01-01", "2001-01-01", "2002-01-01")))
+#'
+#' df$with_columns(
+#'   leap_year = pl$col("date")$dt$is_leap_year()
+#' )
+ExprDT_is_leap_year = function() {
+  .pr$Expr$dt_is_leap_year(self)
 }