+++ src/rust/src/lazy/dsl.rs
@@ -16,8 +16,7 @@ use crate::CONFIG;
 use extendr_api::{extendr, prelude::*, rprintln, Deref, DerefMut};
 use pl::PolarsError as pl_error;
 use pl::{
-    Duration, DurationMethods, IntoSeries, RollingGroupOptions, SetOperation, StringNameSpaceImpl,
-    TemporalMethods,
+    Duration, IntoSeries, RollingGroupOptions, SetOperation, StringNameSpaceImpl, TemporalMethods,
 };
 use polars::lazy::dsl;
 use polars::prelude as pl;
@@ -1469,78 +1468,33 @@ impl RPolarsExpr {
     }
 
     pub fn dt_total_days(&self) -> RResult<Self> {
-        Ok(self
-            .0
-            .clone()
-            .map(
-                |s| Ok(Some(s.duration()?.days().into_series())),
-                pl::GetOutput::from_type(pl::DataType::Int64),
-            )
-            .into())
+        Ok(self.0.clone().dt().total_days().into())
     }
     pub fn dt_total_hours(&self) -> RResult<Self> {
-        Ok(self
-            .0
-            .clone()
-            .map(
-                |s| Ok(Some(s.duration()?.hours().into_series())),
-                pl::GetOutput::from_type(pl::DataType::Int64),
-            )
-            .into())
+        Ok(self.0.clone().dt().total_hours().into())
     }
     pub fn dt_total_minutes(&self) -> RResult<Self> {
-        Ok(self
-            .0
-            .clone()
-            .map(
-                |s| Ok(Some(s.duration()?.minutes().into_series())),
-                pl::GetOutput::from_type(pl::DataType::Int64),
-            )
-            .into())
+        Ok(self.0.clone().dt().total_minutes().into())
     }
     pub fn dt_total_seconds(&self) -> RResult<Self> {
-        Ok(self
-            .0
-            .clone()
-            .map(
-                |s| Ok(Some(s.duration()?.seconds().into_series())),
-                pl::GetOutput::from_type(pl::DataType::Int64),
-            )
-            .into())
+        Ok(self.0.clone().dt().total_seconds().into())
     }
     pub fn dt_total_milliseconds(&self) -> RResult<Self> {
-        Ok(self
-            .0
-            .clone()
-            .map(
-                |s| Ok(Some(s.duration()?.milliseconds().into_series())),
-                pl::GetOutput::from_type(pl::DataType::Int64),
-            )
-            .into())
+        Ok(self.0.clone().dt().total_milliseconds().into())
     }
     pub fn dt_total_microseconds(&self) -> RResult<Self> {
-        Ok(self
-            .0
-            .clone()
-            .map(
-                |s| Ok(Some(s.duration()?.microseconds().into_series())),
-                pl::GetOutput::from_type(pl::DataType::Int64),
-            )
-            .into())
+        Ok(self.0.clone().dt().total_microseconds().into())
     }
     pub fn dt_total_nanoseconds(&self) -> RResult<Self> {
-        Ok(self
-            .0
-            .clone()
-            .map(
-                |s| Ok(Some(s.duration()?.nanoseconds().into_series())),
-                pl::GetOutput::from_type(pl::DataType::Int64),
-            )
-            .into())
+        Ok(self.0.clone().dt().total_nanoseconds().into())
     }
 
     pub fn dt_offset_by(&self, by: Robj) -> RResult<Self> {
         Ok(self.clone().0.dt().offset_by(robj_to!(PLExpr, by)?).into())
+    }
+
+    pub fn dt_is_leap_year(&self) -> Self {
+        self.clone().0.dt().is_leap_year().into()
     }
 
     pub fn repeat_by(&self, by: &RPolarsExpr) -> Self {
@@ -1777,6 +1731,70 @@ impl RPolarsExpr {
 
     pub fn pow(&self, exponent: Robj) -> RResult<Self> {
         Ok(self.0.clone().pow(robj_to!(PLExpr, exponent)?).into())
+    }
+
+    pub fn cut(
+        &self,
+        breaks: Robj,
+        labels: Robj,
+        left_closed: Robj,
+        include_breaks: Robj,
+    ) -> RResult<Self> {
+        let breaks = robj_to!(Vec, f64, breaks)?;
+        let labels = robj_to!(Option, Vec, String, labels)?;
+        let left_closed = robj_to!(bool, left_closed)?;
+        let include_breaks = robj_to!(bool, include_breaks)?;
+        Ok(self
+            .0
+            .clone()
+            .cut(breaks, labels, left_closed, include_breaks)
+            .into())
+    }
+
+    pub fn qcut(
+        &self,
+        probs: Robj,
+        labels: Robj,
+        left_closed: Robj,
+        allow_duplicates: Robj,
+        include_breaks: Robj,
+    ) -> RResult<Self> {
+        let probs = robj_to!(Vec, f64, probs)?;
+        let labels = robj_to!(Option, Vec, String, labels)?;
+        let left_closed = robj_to!(bool, left_closed)?;
+        let allow_duplicates = robj_to!(bool, allow_duplicates)?;
+        let include_breaks = robj_to!(bool, include_breaks)?;
+        Ok(self
+            .0
+            .clone()
+            .qcut(probs, labels, left_closed, allow_duplicates, include_breaks)
+            .into())
+    }
+
+    pub fn qcut_uniform(
+        &self,
+        n_bins: Robj,
+        labels: Robj,
+        left_closed: Robj,
+        allow_duplicates: Robj,
+        include_breaks: Robj,
+    ) -> RResult<Self> {
+        let n_bins = robj_to!(usize, n_bins)?;
+        let labels = robj_to!(Option, Vec, String, labels)?;
+        let left_closed = robj_to!(bool, left_closed)?;
+        let allow_duplicates = robj_to!(bool, allow_duplicates)?;
+        let include_breaks = robj_to!(bool, include_breaks)?;
+        Ok(self
+            .0
+            .clone()
+            .qcut_uniform(
+                n_bins,
+                labels,
+                left_closed,
+                allow_duplicates,
+                include_breaks,
+            )
+            .into())
     }
 
     pub fn over(&self, partition_by: Robj, mapping: Robj) -> RResult<Self> {
@@ -2362,6 +2380,15 @@ impl RPolarsExpr {
             _ => Ok(self.0.clone().str().find(pat, strict).into()),
         }
     }
+
+    fn str_head(&self, n: Robj) -> RResult<Self> {
+        Ok(self.0.clone().str().head(robj_to!(PLExprCol, n)?).into())
+    }
+
+    fn str_tail(&self, n: Robj) -> RResult<Self> {
+        Ok(self.0.clone().str().tail(robj_to!(PLExprCol, n)?).into())
+    }
+
     //binary methods
     pub fn bin_contains(&self, lit: Robj) -> RResult<Self> {
         Ok(self