+++ tests/testthat/test-lazy.R
@@ -31,6 +31,33 @@ test_that("create LazyFrame", {
   )
 })
 
+
+test_that("LazyFrame serialize/deseialize", {
+  skip_if_not_installed("jsonlite")
+
+  df = pl$DataFrame(
+    a = 1:3,
+    b = letters[1:3]
+  )
+
+  lf = df$lazy()$filter(pl$col("a") >= 2)$select("b")
+  json = lf$serialize()
+
+  expect_snapshot(jsonlite::prettify(json))
+
+  expect_true(lf$collect()$equals(
+    pl$deserialize_lf(json)$collect()
+  ))
+
+  expect_grepl_error(
+    df$lazy()$select(
+      pl$col("a")$map_elements(\(x) -abs(x))
+    )$serialize(),
+    "serialize not supported for this 'opaque' function"
+  )
+})
+
+
 test_that("LazyFrame, custom schema", {
   df = pl$LazyFrame(
     iris,